Fix TestEnvironmentOfChildProcess to handle newlines
Something changed recently in the CI environment, such that Jenkins is now putting PR/commit/comment information in environment variables.  This is breaking the TestEnvironmentOfChildProcess test, which assumes (poorly) environment variables won't have newlines in them, as it's using newlines for parsing.  This fix basically just switches to using something other than a newline for parsing.  It also includes a few other simplifications.