Don't copy WP8 aseemblies to test overlay
Before we moved to using packer to slim down our test drop, we had
Jenkins exclude the System.Threading.Tasks.Dataflow.WP8 assemblies when
uploading build artifacts. This meant that the OSX and Linux jobs would
never copy them down and hence they could not be picked up by
run-test.sh (and overwrite the correct version of
System.Threading.Tasks.Dataflow when building the test overlay).

With packer, we now archive everything, so depending on the order in
which find discovered files, we may have copied over the WP8 version,
which does not have ETW support.

For now, update run-test.sh to exclude this folder when copying stuff
over. The correct long term fix is to just do what we do on
Windows (have the build produce the correct test layouts and get out of
this run-test.sh test overlay game).

Fixes #3520