Write a message when an xUnit invocation fails
Previously finding out which xunit projects failed was difficult because
unlike Windows we didn't write a message when xUnit returned with a non
zero exit code.  In cases where tests failed, it was somewhat easy to
search the logs for xUnits "a test failed" messages but if the runtime
crashed, things were much harder.

Update run-test.sh to write an error message in the same style as what
we do on windows as part of the xUnit invocation from MSBuild.

While I was in the area, I reduced some noise in the log by pushing the
output of a grep for supported platforms to /dev/null. We print a
message anyway when we detect a project is unsupported, so the output of
grep is not interesting.