Renable some HttpWebResponse tests
Modify tests checking for presence/absence of 'Content-Type' response
headers. Switch to using the new HttpTestServers2 class which supports
these kind of endpoints.

Simplify the tests to be [Fact] and not [Theory]. These tests don't require
to be run against both HTTP and HTTPS servers. I will be making these
kinds of changes further in order to streamline the number of tests
running in CI.

This PR also fixes #2385.