Re-enable some WinHttpHandler unit tests
Turned on tests:

SendAsync_NoWinHttpDecompressionSupportAndResponseBodyIsDeflateCompressed_ExpectedResponse
SendAsync_NoWinHttpDecompressionSupportAndResponseBodyIsGZipCompressed_ExpectedResponse

since the issue with clrcompression.dll was resolved a while ago.

Removed a *Nix annotation suppressing a unit test. This is not needed in
the test sources since the WinHttpHandler library and tests are only
supported on Windows and won't run anymore in CI (disabled in CSPROJ
files) for non-Windows platforms.