Fix WinHttpHandler on Win7
When WinHttpHandler was recently switched to use async WinHTTP pattern, it started setting an option which improves performance. However, the WINHTTP_OPTION_ASSURED_NON_BLOCKING_CALLBACKS option is not available on downlevel Windows versions. This was causing an exception to be thrown when sending an HTTP request.

The fix is to ignore the ERROR_WINHTTP_INVALID_OPTION error code when trying to set this option.