Validate linger time.
Some platforms behave strangely when confronted with linger times that
are out-of-range. In particular, OS X silently truncates
`linger::l_linger` to 16 bits and interprets the results as a `short`.
This change validates the linger time before calling `setsockopt` to
ensure reliable behavior across platforms.

On OS X, the maximum linger time is actually in units of ticks, so
validation is a bit more involved than on other platforms (where the
max time is defined by the size of the integer used to store the
value): on the first call to `SetLingerOption`, the number of ticks
per second is obtained via `sysconf()` and is then used to derive
the maximum linger time by simple division. This does mean that the
maximum linger time on OS X may vary with the number of ticks per
second, and so may vary between machines.