Fix CurlHandler deadlock with synchronous CopyToAsync
CurlHandler calls to HttpContext.CopyToAsync with a stream that CopyToAsync should write to asynchronously.  If, however, the call to CopyToAsync writes to it synchronously as part of the synchronous call to CopyToAsync, CurlHandler will deadlock.

As a precaution, we ensure that the CopyToAsync call is asynchronous from the perspective of CurlHandler.