Update a socket test to accommodate differences on *nix.
On *nix, the SendTo APIs return ENETUNREACH instead of EFAULT when confronted
with a socket address with an invalid address family: these platforms check
the family of the provided socket address before checking its size (as long as
the socket address is large enough to store an address family).

Fixes #4003.