System.Net.NetworkInformation 4.1 contract support for N (UWP) and K
This change allows the System.Net.NetworkInformation 4.1 contract to be properly supported with the "generations" for the NuGet packages. This means that the 4.1.* NuGet package will be supported on both K (DNXCORE50) and N (NETCORE50) targets.

Except for a few methods that already worked for N (UWP) in the 4.0 contract, the rest are stubbed out and throw PlatformNotSupportedException on N.

It is still to be decided whether the full 4.1 contract for N (UWP) will continue to use the WinRT APIs for the implementation or switch over to common sources used in K. That will depend on whether the dependent Win32 APIs will pass the WACK.

[tfs-changeset: 1541417]