Pass extra arguments to MSBuild, not make
CoreFX folks are going to want to be able to do things like build.sh
/p:SkipTests=true but today all unrecognized arguments to build.sh are
passed to make instead of MSBuild.

This is a bad default for us.

If we have the need to start passing arguments to build the native build
process and managed, we should agument build.sh to do something like
treating all unrecgonized arguments after the "native" token to be
things to pass to make and have the default behavior do the right thing
most of the time which is pass arguments to MSBuild.