Finish support for X509BasicConstraintsExtension on Unix
The form of the extension called szOID_BASIC_CONSTRAINTS (2.5.29.10) doesn't have a public documented ASN structure, and is legacy.  Rather than add code to support it, and tests to verify it, just make the Unix implementation treat it as an unknown extension object.  But, in the event someone manually creates an X509BasicConstraintsExtension object with the legacy OID, throw a PlatformNotSupportedException on decode.

The encoding of the extension (OID 2.5.29.19, from ITU X.509 03/2000) is performed with the fledgling DerEncoder.