Fixing tests having different behavior
We discovered some tests were having a different behavior due to the
nondeterministic nature of calling a disposal of a socket that is doing
an EndAccept operation in a different thread.

Also, the test was incorrectly propagating the expected SocketException
coming from the client socket, using the server as an intermediary.  Now
it's correctly done using a SocketError.

Fix #4052

Fixing race condition

Added a handler so the connect is correctly awaited.

Format Fix

Added some blank lines required by coding guidelines