Handle CURLE_UNKNOWN_OPTION when registering HTTPS callbacks on older versions of curl.
Prior to curl 7.39 the case label for CURLOPT_SSL_CTX_FUNCTION was compiled out, resulting in CURLE_UNKNOWN_OPTION.  In 7.39 they changed it to always present, but reporting CURLE_NOT_BUILT_IN when the backend SSL provider didn't support that callback.

The code was written based on the behavior of curl 7.43, so didn't know about the legacy behavioral difference.