Address // PERF markers in NetworkInformation
There was an additional marker regarding using sysctl to query the machine's address count. I don't think it's actually worth bothering, since all we are doing instead is adding up the number of addresses from the list of interfaces we already have. It's probably faster than PInvoking for a single number.